💄(frontend) create IconBG component
Create IconBG component, a premade Icon component
with a background color.
