+++ src/frontend/apps/impress/src/components/index.ts
@@ -2,7 +2,8 @@ export * from './Box';
 export * from './BoxButton';
 export * from './Card';
 export * from './DropButton';
-export * from './IconOptions';
+export * from './Icon';
+export * from './InfiniteScroll';
 export * from './Link';
 export * from './Text';
 export * from './TextErrors';
